+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Improved ChatMessage to show token image, name, and user. [#2608](https://github.com/dmdorman/hero6e-foundryvtt/issues/2608)
 - Martial FLASH now properly targets the sense group.
+- FLASH for normal senses are implemented as if they target the entire sense group.
 
 ## Version 4.1.12 20250813